[IMP] nirun_service: Update
- Count Active service request as Participant
- Lock main field on Service Request after approve
- User can delete Service request
- Fix service time bug - no store minute
- Service Request list view action approve and complete
- Can create Service Request in bulk
- Full State control for Service Request
- Auto set period_end on Revoke or Complete if not set or in future
- State button on Patient & Encounter Form

+++ nirun_service/models/service_time.py
@@ -18,8 +18,8 @@ class HealthcareServiceAvailableTime(models.Model):
         "ni.timing.dow", "ni_service_time_dow", "timing_id", "dow_id"
     )
     all_day = fields.Boolean()
-    start_time = fields.Integer()
-    end_time = fields.Integer()
+    start_time = fields.Float()
+    end_time = fields.Float()
 
     display_day = fields.Char("Day", compute="_compute_dow_txt")
     display_time = fields.Char("Time", compute="_compute_time_txt")
